VirtualBoxImpl.cpp: size_t/DWORD warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@18484 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/VirtualBoxImpl.cpp
@@ -4174,7 +4174,7 @@ DECLCALLBACK(int) VirtualBox::ClientWatcher (RTTHREAD /* thread */, void *pvUser
             /* release the caller to let uninit() ever proceed */
             autoCaller.release();
 
-            DWORD rc = ::WaitForMultipleObjects (1 + cnt + cntSpawned,
+            DWORD rc = ::WaitForMultipleObjects ((DWORD)(1 + cnt + cntSpawned),
                                                  handles, FALSE, INFINITE);
 
             /* Restore the caller before using VirtualBox. If it fails, this
